Adds signature checking for mock spec object method calls
Mock can accept a spec object / class as argument, making sure that
accessing attributes that do not exist in the spec will cause an
AttributeError to be raised, but there is no guarantee that the spec's
methods signatures are respected in any way. This creates the possibility
to have faulty code with passing unittests and assertions.

Passes the spec's attribute with the same name to the child mock (spec-ing
the child).
Sets _mock_check_sig if the given spec is callable.
Adds unit tests to validate the fact that the spec method signatures are
respected.

mock_open creates a Mock object with the built-in open function as a spec,
thus, open's signature has to be respected. There are a few unit tests
that does not respect open's signature, meaning that they fail under this
patch, if not fixed.

Fixes the mentioned unit tests.