consumes self/cls arguments when patching methods
Currently, when patching methods with autospec, their self / cls
arguments are not consumed, causing call asserts to fail (they
expect an instance / class reference as the first argument).

When patching with autospec, a Mock object is created with the
given spec. Due to commit 47e8f38c11, the method signature will
be checked for Mock objects with spec objects. Thus, there is
no need to add another signature check on create_autospec.