Checks autospeced methods argument orders
Currently, mock only checks the validity of signatures through
binding (this can guarantee that the argument count is correct,
and that kwargs are properly passed-in), but there is no guarantee
that the arguments are in the right order. For example, if the function:

def foo(a, b, c):
    pass

is mocked and spec'ed, calls like: mock_foo(b, a, d) are still
considered "valid" from mock's point of view.

One way to partially address this issue is by using sentinels as
arguments, and checking if the calls pass them in the *right* order.
No assertions can be made if the arguments are not sentinels.

Thus, calls like this will raise TypeError exceptions:

mock_foo(sentinel.b, sentinel.a, sentinel.d)